Split concrete types out for Enum and Class TypeCompilettes
