+++ ulox/ulox.core/Package/Runtime/Compiler/Compilettes/TypeInitCompilette.cs
@@ -10,14 +10,10 @@
         
         public void Process(Compiler compiler)
         {
-            var initName = TypeCompilette.InitMethodName.String;
+            var initName = ClassTypeCompilette.InitMethodName.String;
             byte constant = compiler.AddCustomStringConstant(initName);
             compiler.Function(initName, FunctionType.Init);
             compiler.EmitPacket(new ByteCodePacket(OpCode.METHOD, constant,0,0));
-        }
-
-        public void Start(TypeCompilette typeCompilette)
-        {
         }
     }
 }
