+++ ulox/ulox.core/Package/Runtime/Interfaces/ITypeBodyCompilette.cs
@@ -3,7 +3,5 @@
     public interface ITypeBodyCompilette : ICompilette
     {
         TypeCompiletteStage Stage { get; }
-
-        void Start(TypeCompilette typeCompilette);
     }
 }
